+++ static/js/browser2/observation_chart_section.tsx
